+++ hub/core/sample.py
@@ -114,30 +114,3 @@ class Sample:
 
     def __repr__(self):
         return str(self)
-
-
-def load(path: str) -> Sample:
-    """Utility that loads data from a file into a `np.ndarray` in 1 line of code. Also provides access to all important metadata.
-
-    Note:
-        No data is actually loaded until you try to read a property of the returned `Sample`. This is useful for passing along to
-            `tensor.append` and `tensor.extend`.
-
-    Examples:
-        >>> sample = hub.load("path/to/cat.jpeg")
-        >>> type(sample.array)
-        <class 'numpy.ndarray'>
-        >>> sample.compression
-        'jpeg'
-
-    Supported File Types:
-        image: png, jpeg, and all others supported by `PIL`: https://pillow.readthedocs.io/en/stable/handbook/image-file-formats.html#fully-supported-formats
-
-    Args:
-        path (str): Path to a supported file.
-
-    Returns:
-        Sample: Sample object. Call `sample.array` to get the `np.ndarray`.
-    """
-
-    return Sample(path)
